requirements.txt
# Core Dependencies
playwright==1.42.0
fake-useragent==1.4.0
python-dotenv==1.0.0

# AI & Data Processing
openai==1.12.0
anthropic==0.18.1

# Data Analysis
pandas==2.2.0
numpy==1.26.4

# Visualization
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
